requirements/test.txt
-r dev.txt

httpx>=0.26.0
pytest-mock>=3.12.0
